third_party/WebKit/LayoutTests/platform/mac/plugins/overlay-scrollbar-mouse-capture-expected.txt


This tests whether scrolling still works correctly when an overlay scrollbar is over a plugin. The plugin should still receive mouse down/up events when clicking an overlay scrollbar. Scrolling should still work correctly too. However mouse capture should not be started on the plugin as this would interfere with events going to the scrollbar.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL eventHistory.length should be 2. Was 3.
PASS eventHistory[0] is "plugin.mousedown"
FAIL eventHistory[1] should be plugin.mouseup. Was plugin.mousemove.
FAIL window.scrollY should not be 0.
PASS successfullyParsed is true

TEST COMPLETE
